- Shopping list refactored, basic error handling for shopping list and ingredient

+++ code/app/src/main/java/com/example/snackoverflow/AddIngredientFragment.java
@@ -22,6 +22,8 @@ import androidx.fragment.app.DialogFragment;
 
 import com.google.android.material.textfield.TextInputLayout;
 
+import org.w3c.dom.Text;
+
 import java.text.DateFormat;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
@@ -38,9 +40,8 @@ import java.util.Objects;
  * @see Ingredient
  * */
 public class AddIngredientFragment extends DialogFragment {
-    // declaring the variables for Fields in the XML file
-    private TextInputLayout ingredientDescLayout;
-    private TextInputLayout ingredientCategoryLayout;
+
+    private TextInputLayout ingredientDescLayout, ingredientCategoryLayout, ingredientUnitLayout, ingredientDateLayout, ingredientAmountLayout;
     private EditText ingredientDesc;
     private EditText ingredientAmount;
     private EditText ingredientUnit;
@@ -48,6 +49,8 @@ public class AddIngredientFragment extends DialogFragment {
     private RadioButton ingredientLocation;
     private RadioGroup locationRadioGroup;
     private EditText ingredientCategory;
+    private String ingredientLocationString;
+    private boolean isNull = false;
 
     /**
      * For Starting the fragment.
@@ -125,11 +128,13 @@ public class AddIngredientFragment extends DialogFragment {
 
             @Override
             public void afterTextChanged(Editable s) {
+
                 // Set the error if the chaarcter count is greater the max length allowed.
                 if (s.length() > layout.getCounterMaxLength()) {
                     layout.setError("Max character length is " + layout.getCounterMaxLength());
-                } else {
-                    layout.setError(null);
+                }
+                else if (s.length() == 0) {
+                    layout.setError("Title cannot be empty");
                 }
             }
         });
@@ -188,20 +193,48 @@ public class AddIngredientFragment extends DialogFragment {
      */
     public Ingredient parseFilledValues(View view) {
         DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+        Integer ingredientAmountInt, ingredientUnitInt;
         String ingredientDescString = ingredientDesc.getText().toString();
-        Integer ingredientAmountInt = (int) Double.parseDouble(ingredientAmount.getText().toString());
-        Integer ingredientUnitInt = Integer.parseInt(ingredientUnit.getText().toString());
+        try {
+            ingredientAmountInt = (int) Double.parseDouble(ingredientAmount.getText().toString());
+        } catch (NumberFormatException e) {
+            ingredientAmountInt = 0;
+            isNull = true;
+        }
+        try {
+            ingredientUnitInt = Integer.parseInt(ingredientUnit.getText().toString());
+        } catch (NumberFormatException e) {
+            ingredientUnitInt = 0;
+            isNull = true;
+        }
+
         String ingredientBestBeforeString = ingredientBestBefore.getText().toString();
-        String ingredientCategoryString = ingredientCategory.getText().toString();
 
-        ingredientLocation = (RadioButton) view.findViewById(locationRadioGroup.getCheckedRadioButtonId());
-        String ingredientLocationString = ingredientLocation.getText().toString();
-
+        if (ingredientBestBeforeString.length() == 0) {
+            isNull = true;
+        }
         Date ingredientBestBeforeDate = null;
         try {
             ingredientBestBeforeDate = dateFormat.parse(ingredientBestBeforeString);
         } catch (ParseException e) {
+            isNull = true;
             e.printStackTrace();
+        }
+
+        int selectedLocation = locationRadioGroup.getCheckedRadioButtonId();
+        ingredientLocation = (RadioButton) view.findViewById(selectedLocation);
+        String ingredientLocationString;
+        try {
+            ingredientLocationString = ingredientLocation.getText().toString();
+        } catch (NullPointerException e) {
+            ingredientLocationString = "";
+            isNull = true;
+        }
+
+        String ingredientCategoryString = ingredientCategory.getText().toString();
+
+        if (ingredientCategoryString.length() == 0) {
+            isNull = true;
         }
         // return the new ingredient created from the values inputted by the user
         return new Ingredient(ingredientDescString, ingredientBestBeforeDate, ingredientLocationString, ingredientAmountInt, ingredientUnitInt, ingredientCategoryString);
@@ -219,6 +252,9 @@ public class AddIngredientFragment extends DialogFragment {
         View view = LayoutInflater.from(getActivity()).inflate(R.layout.add_ingredient_fragment_layout, null);
         ingredientDescLayout = view.findViewById(R.id.ingredient_description_layout);
         ingredientCategoryLayout = view.findViewById(R.id.ingredient_category_layout);
+        ingredientUnitLayout = view.findViewById(R.id.ingredient_unit_layout);
+        ingredientDateLayout = view.findViewById(R.id.ingredient_bestBefore_layout);
+        ingredientAmountLayout = view.findViewById(R.id.ingredient_amount_layout);
         ingredientDesc = view.findViewById(R.id.ingredient_description_editText);
         ingredientBestBefore = view.findViewById(R.id.ingredient_bestBefore_editText);
         locationRadioGroup = view.findViewById(R.id.ingredient_location_radioGroup);
@@ -247,8 +283,9 @@ public class AddIngredientFragment extends DialogFragment {
                         @Override
                         public void onClick(DialogInterface dialog, int which) {
                             Ingredient newIngredient = parseFilledValues(view);
-                            // Add the ingredient to Firestore Database
-                            FirestoreDatabase.addIngredient(newIngredient);
+                            if (!isNull) {
+                                FirestoreDatabase.addIngredient(newIngredient);
+                            }
                         }
                     }).create();
         } else {
@@ -272,24 +309,26 @@ public class AddIngredientFragment extends DialogFragment {
                     .setPositiveButton("OK", new DialogInterface.OnClickListener() {
                         @Override
                         public void onClick(DialogInterface dialog, int which) {
-                            // Parse the values filled by the user
+
                             Ingredient updatedIngredient = parseFilledValues(view);
 
-                            // Modify initial ingredient
-                            initialIngredient.setTitle(updatedIngredient.getTitle());
-                            initialIngredient.setAmount(updatedIngredient.getAmount());
-                            initialIngredient.setBestBefore(updatedIngredient.getBestBefore());
-                            initialIngredient.setUnit(updatedIngredient.getUnit());
-                            initialIngredient.setCategory(updatedIngredient.getCategory());
-                            initialIngredient.setLocation(updatedIngredient.getLocation());
-                            if (isShoppingListItem) {
-                                // If the item is from shopping list, add it to the database
-                                FirestoreDatabase.addIngredient(initialIngredient);
-                                FirestoreDatabase.deleteShoppingItem(initialIngredient.getTitle());
-                                ((ShoppingListActivity) getActivity()).removeIngredient();
-                            } else {
-                                // Modify the existing ingredient in the database
-                                FirestoreDatabase.modifyIngredient(initialIngredient);
+                            if (!isNull) {
+                                // Modify initial ingredient
+                                initialIngredient.setTitle(updatedIngredient.getTitle());
+                                initialIngredient.setAmount(updatedIngredient.getAmount());
+                                initialIngredient.setBestBefore(updatedIngredient.getBestBefore());
+                                initialIngredient.setUnit(updatedIngredient.getUnit());
+                                initialIngredient.setCategory(updatedIngredient.getCategory());
+                                initialIngredient.setLocation(updatedIngredient.getLocation());
+                                if (isShoppingListItem) {
+                                    // If the item is from shopping list, add it to the database
+                                    FirestoreDatabase.addIngredient(initialIngredient);
+                                    FirestoreDatabase.deleteShoppingItem(initialIngredient.getTitle());
+                                    ((ShoppingListActivity) getActivity()).removeIngredient();
+                                } else {
+                                    // Modify the existing ingredient in the database
+                                    FirestoreDatabase.modifyIngredient(initialIngredient);
+                                }
                             }
                         }
                     }).create();
